refactor: 자동차 상태 기능 클래스명 수정
기능의 역할에 맞게 클래스 및 변수명 수정 및 신규 테스트 케이스 추가

* class
 * CarStatus -> CarProgress
* variable
  * progress -> count

* Test Case
  * 자동차상태 - 상태바 초기화 상태 테스트
